+++ db_test/db_mac_tests/yas_db_result_set_tests.mm
@@ -122,13 +122,13 @@
     auto data_value = result_set.column_value("data_field");
     auto null_value = result_set.column_value("null_field");
 
-    XCTAssertTrue(int_value.type() == typeid(yas::db::int64));
+    XCTAssertTrue(int_value.type() == typeid(yas::db::integer));
     XCTAssertTrue(float_value.type() == typeid(yas::db::float64));
     XCTAssertTrue(string_value.type() == typeid(yas::db::string));
     XCTAssertTrue(data_value.type() == typeid(yas::db::blob));
     XCTAssertTrue(null_value.type() == typeid(yas::db::null));
 
-    XCTAssertEqual(int_value.value<yas::db::int64>(), 1);
+    XCTAssertEqual(int_value.value<yas::db::integer>(), 1);
     XCTAssertEqual(float_value.value<yas::db::float64>(), 2.0);
     XCTAssertEqual(string_value.value<yas::db::string>(), "string_value");
 
@@ -179,13 +179,13 @@
     auto &data_value = map.at("data_field");
     auto &null_value = map.at("null_field");
 
-    XCTAssertTrue(int_value.type() == typeid(yas::db::int64));
+    XCTAssertTrue(int_value.type() == typeid(yas::db::integer));
     XCTAssertTrue(float_value.type() == typeid(yas::db::float64));
     XCTAssertTrue(string_value.type() == typeid(yas::db::string));
     XCTAssertTrue(data_value.type() == typeid(yas::db::blob));
     XCTAssertTrue(null_value.type() == typeid(yas::db::null));
 
-    XCTAssertEqual(int_value.value<yas::db::int64>(), 1);
+    XCTAssertEqual(int_value.value<yas::db::integer>(), 1);
     XCTAssertEqual(float_value.value<yas::db::float64>(), 2.0);
     XCTAssertEqual(string_value.value<yas::db::string>(), "string_value");
